small changes to user auth

+++ Back-End/routes/userRoutes.js
@@ -4,10 +4,10 @@ const Post = require("../models/Post");
 const User = require("../models/User");
 
 userRoutes.post("/article/post", async (req, res) => {
-  if (req.isAuthenticated()) {
+  if (req.isAuthenticated() && req.user) {
     const postObj = await Post.create(req.body);
     await User.updateOne(
-      { _id: postObj.author },
+      { _id: req.user._id },
       { $push: { createdPosts: postObj._id } }
     );
     res.json(postObj);
@@ -17,12 +17,12 @@ userRoutes.post("/article/post", async (req, res) => {
 });
 
 userRoutes.post("/article/delete", async (req, res) => {
-  if (req.isAuthenticated()) {
+  if (req.isAuthenticated() && req.user) {
     try {
-      let { post_id, user_id } = req.body;
+      let { post_id } = req.body;
       await Post.deleteOne({ _id: post_id });
       await User.updateOne(
-        { _id: user_id },
+        { _id: req.user._id },
         { $pull: { createdPosts: post_id } }
       );
       res.json({ error: false });
